+++ app_dart/lib/cocoon_service.dart
@@ -29,6 +29,8 @@ export 'src/request_handlers/github_rate_limit_status.dart';
 export 'src/request_handlers/github_webhook.dart';
 export 'src/request_handlers/postsubmit_luci_subscription.dart';
 export 'src/request_handlers/presubmit_luci_subscription.dart';
+export 'src/request_handlers/presubmit_luci_subscription_ordered.dart';
+export 'src/request_handlers/presubmit_luci_subscription_unordered.dart';
 export 'src/request_handlers/push_build_status_to_github.dart';
 export 'src/request_handlers/push_gold_status_to_github.dart';
 export 'src/request_handlers/readiness_check.dart';
